Backend/fixngo-backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
boto3==1.35.84
botocore==1.35.84
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.3
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
eventlet==0.38.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-engineio==4.11.2
python-socketio==5.12.1
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
service-identity==24.2.0
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
